Fixup FQNs in javadocs and move StatisticsConfig
