+++ geode-core/src/main/java/org/apache/geode/distributed/internal/DistributionConfig.java
@@ -208,7 +208,7 @@ import org.apache.geode.internal.security.SecurableCommunicationChannel;
 import org.apache.geode.internal.tcp.Connection;
 import org.apache.geode.logging.internal.spi.LogConfig;
 import org.apache.geode.logging.internal.spi.LogWriterLevel;
-import org.apache.geode.statistics.StatisticsConfig;
+import org.apache.geode.internal.statistics.StatisticsConfig;
 
 /**
  * Provides accessor (and in some cases mutator) methods for the various GemFire distribution
