+++ geode-core/src/main/java/org/apache/geode/distributed/internal/InternalLocator.java
@@ -99,7 +99,7 @@ import org.apache.geode.management.internal.configuration.messages.ClusterManage
 import org.apache.geode.management.internal.configuration.messages.SharedConfigurationStatusRequest;
 import org.apache.geode.management.internal.configuration.messages.SharedConfigurationStatusResponse;
 import org.apache.geode.security.AuthTokenEnabledComponents;
-import org.apache.geode.statistics.StatisticsConfig;
+import org.apache.geode.internal.statistics.StatisticsConfig;
 
 /**
  * Provides the implementation of a distribution {@code Locator} as well as internal-only
