+++ geode-core/src/main/java/org/apache/geode/internal/logging/ManagerLogWriterFactory.java
@@ -31,7 +31,7 @@ import org.apache.geode.internal.OSProcess;
 import org.apache.geode.internal.process.ProcessLauncherContext;
 import org.apache.geode.internal.util.LogFileUtils;
 import org.apache.geode.logging.internal.spi.LogConfig;
-import org.apache.geode.statistics.StatisticsConfig;
+import org.apache.geode.internal.statistics.StatisticsConfig;
 
 /**
  * Factory for creating a {@link ManagerLogWriter}.
