+++ geode-core/src/main/java/org/apache/geode/internal/statistics/GemFireStatSampler.java
@@ -39,7 +39,6 @@ import org.apache.geode.internal.logging.log4j.LogMarker;
 import org.apache.geode.internal.statistics.platform.OsStatisticsFactory;
 import org.apache.geode.internal.statistics.platform.ProcessStats;
 import org.apache.geode.logging.internal.spi.LogFile;
-import org.apache.geode.statistics.StatisticsConfig;
 
 /**
  * GemFireStatSampler adds listeners and rolling archives to HostStatSampler.
